requirements.txt
astroid==2.15.6
attrs==23.1.0
black==23.7.0
build==0.10.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.7
cloudscraper==1.2.71
crashtest==0.4.1
cryptography==41.0.3
cssselect==1.2.0
cssutils==2.7.1
deepdiff==6.3.1
dill==0.3.7
distlib==0.3.7
dulwich==0.21.6
filelock==3.12.3
firebase-admin==6.2.0
free-proxy==1.1.1
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
httplib2==0.22.0
identify==2.5.27
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
isort==5.12.0
jaraco.classes==3.3.0
jeepney==0.8.0
jsonschema==4.17.3
keyring==24.2.0
lazy-object-proxy==1.9.0
lxml==4.9.3
mccabe==0.7.0
more-itertools==10.1.0
msgpack==1.0.5
mypy==1.5.1
mypy-extensions==1.0.0
nodeenv==1.8.0
ordered-set==4.1.0
packaging==23.1
pathspec==0.11.2
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
pre-commit==3.3.3
premailer==3.10.0
proto-plus==1.22.3
protobuf==4.24.1
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pylint==2.17.5
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pyshorteners==1.0.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.3
six==1.16.0
termcolor==2.3.0
tgtg==0.16.0
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.8.7
types-pytz==2023.3.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.4
wrapt==1.15.0
yagmail==0.15.293
yaspin==3.0.0
zipp==3.16.2
